[2702] - Make geocoding more accurate
- Providers need to have 'provider_name' included in geocoding
- Sites need to have 'location_name' included in geocoding
- This will improve the accuracy of lat / long coordinates
